fix: move shared types/functions out of "use client" module
Move CitationQuote, CitationHealthSummary types and computeCitationHealth()
from the "use client" CitationOverlay component to the server-compatible
citation-data.ts module. Next.js cannot call client-exported functions from
server components during static generation.
