+++ apps/web/src/components/wiki/CitationOverlay.tsx
@@ -11,33 +11,7 @@ import {
   Clock,
   ExternalLink,
 } from "lucide-react";
-
-/** Citation quote data from the wiki-server API */
-export interface CitationQuote {
-  footnote: number;
-  url: string | null;
-  claimText: string;
-  sourceQuote: string | null;
-  sourceTitle: string | null;
-  quoteVerified: boolean;
-  verificationScore: number | null;
-  verifiedAt: string | null;
-  accuracyVerdict: string | null;
-  accuracyScore: number | null;
-  accuracyIssues: string | null;
-  accuracyCheckedAt: string | null;
-}
-
-/** Summary stats for page-level banner */
-export interface CitationHealthSummary {
-  total: number;
-  verified: number;
-  accurate: number;
-  inaccurate: number;
-  unsupported: number;
-  minorIssues: number;
-  unchecked: number;
-}
+import type { CitationQuote } from "@/lib/citation-data";
 
 interface VerdictConfig {
   icon: typeof CheckCircle2;
@@ -331,49 +305,3 @@ export function CitationOverlay({ quotes }: { quotes: CitationQuote[] }) {
   );
 }
 
-/**
- * Computes a summary of citation health from the quotes array.
- */
-export function computeCitationHealth(
-  quotes: CitationQuote[]
-): CitationHealthSummary {
-  let verified = 0;
-  let accurate = 0;
-  let inaccurate = 0;
-  let unsupported = 0;
-  let minorIssues = 0;
-  let unchecked = 0;
-
-  for (const q of quotes) {
-    if (q.accuracyVerdict) {
-      switch (q.accuracyVerdict) {
-        case "accurate":
-          accurate++;
-          break;
-        case "inaccurate":
-          inaccurate++;
-          break;
-        case "unsupported":
-          unsupported++;
-          break;
-        case "minor_issues":
-          minorIssues++;
-          break;
-      }
-    } else if (q.quoteVerified) {
-      verified++;
-    } else {
-      unchecked++;
-    }
-  }
-
-  return {
-    total: quotes.length,
-    verified,
-    accurate,
-    inaccurate,
-    unsupported,
-    minorIssues,
-    unchecked,
-  };
-}
